% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggleoNT.R
\name{ggleoNT}
\alias{ggleoNT}
\title{Title}
\usage{
ggleoNT(f, x, tol, N)
}
\arguments{
\item{f}{the function with any variables}

\item{x}{the guessing start point}

\item{tol}{Accuracy ex(0.001)}

\item{N}{The maximum number of iterations}
}
\value{
the maximum point
}
\description{
Title
}
\examples{
library(ggleo)
ggleoNT(f,x,tol,N)
}
